requirements.txt
Django==6.0.3
djangorestframework==3.17.1
psycopg2-binary==2.9.11
gunicorn>=21.2.0,<23

